% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatHashes}
\alias{formatHashes}
\title{formatHashes}
\usage{
formatHashes(x, units = "auto", ...)
}
\arguments{
\item{x}{a number (of hashes)}

\item{units}{what units you want the hashes displayed in
(e.g. "kH", "MH", etc.").  By default, the function will automatically
try to figure out the best unit.}

\item{...}{other arguments (not currently in use)}
}
\value{
string
}
\description{
pretty print for hash rate
}
\details{
\itemize{
  \item 1 kH/s is 1,000 (one thousand) hashes per second.   
  \item 1 MH/s is 1,000,000 (one million) hashes per second.   
  \item 1 GH/s is 1,000,000,000 (one billion) hashes per second.   
  \item 1 TH/s is 1,000,000,000,000 (one trillion) hashes per second.   
  \item 1 PH/s is 1,000,000,000,000,000 (one quadrillion) hashes per second.   
  \item 1 EH/s is 1,000,000,000,000,000,000 (one quintillion) hashes per second.  
}
}
\examples{
\dontrun{
formatHashes(getNetHash())
}
}
\references{
\url{http://bitcoin.stackexchange.com/a/21498}
}
\author{
Garrett See
}
