% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{getAddress}
\alias{getAddress}
\title{getAddress}
\usage{
getAddress(userpwd = getOption("monero.wallet.userpwd"),
  ip = getOption("monero.wallet.ip", "127.0.0.1"),
  port = getOption("monero.wallet.port", "18082"))
}
\arguments{
\item{userpwd}{user and password for monero-wallet-rpc calls. 
(e.g. "username:password")}

\item{ip}{ip address of wallet}

\item{port}{port of wallet}
}
\value{
chr address
}
\description{
get the wallet's address
}
\author{
Garrett See
}
