% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallet.R
\name{getBalance}
\alias{getBalance}
\title{getBalance}
\usage{
getBalance(userpwd = getOption("monero.wallet.userpwd"),
  ip = getOption("monero.wallet.ip", "127.0.0.1"),
  port = getOption("monero.wallet.port", "18082"))
}
\arguments{
\item{userpwd}{user and password for monero-wallet-rpc calls. 
(e.g. "username:password")}

\item{ip}{ip address of wallet}

\item{port}{port of wallet}
}
\value{
\itemize{
 \item balance unsigned int; The total balance of the current simplewallet in
 session.
 \item unlocked_balance unsigned int; Unlocked funds are those funds that are
 sufficiently deep enough in the Monero blockchain to be considered safe to
 spend.
}
}
\description{
get the wallet's balance.
}
\author{
Garrett See
}
