% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monero.R
\name{getBlock}
\alias{getBlock}
\title{getBlock}
\usage{
getBlock(height, hash, ip = getOption("monerod.ip", "127.0.0.1"),
  port = getOption("monerod.port", 18081))
}
\arguments{
\item{height}{int; the block's height.}

\item{hash}{chr; the block's hash.}

\item{ip}{daemon ip address}

\item{port}{daemon port}
}
\value{
\itemize{
 \item blob - string; Hexadecimal blob of block information.
 \item block_header - A structure containing block header information. See getlastblockheader.
  \itemize{
    \item block_size
    \item depth
    \item difficulty
    \item hash
    \item height
    \item major_version
    \item minero_version
    \item nonce
    \item num_txes
    \item orphan_status
    \item prev_hash
    \item reward
    \item timestamp
  }
 \item json - json string; JSON formatted block details:
  \itemize{
  \item major_version - Same as in block header.
  \item minor_version - Same as in block header.
  \item timestamp - Same as in block header.
  \item prev_id - Same as prev_hash in block header.
  \item nonce - Same as in block header.
  \item miner_tx - Miner transaction information
    \itemize{
       \item version - Transaction version number.
       \item unlock_time - The block height when the coinbase transaction becomes spendable.
       \item vin - List of transaction inputs:
        \itemize{
         \item gen - Miner txs are coinbase txs, or "gen".  height - This block height, a.k.a. when the coinbase is generated.
        }
       \item vout - List of transaction outputs. Each output contains:
        \itemize{
          \item amount - The amount of the output, in atomic units.
          \item target -
          \item key -
        }
       \item extra - Usually called the "transaction ID" but can be used to include any random 32 byte/64 character hex string.
       \item signatures - Contain signatures of tx signers. Coinbased txs do not have signatures.
     }
   \item tx_hashes - List of hashes of non-coinbase transactions in the block. If there are no other transactions, this will be an empty list.
  }
  \item status - string; General RPC error code. "OK" means everything looks good.
}
}
\description{
Full block information can be retrieved by either block height or hash.
}
\details{
Provide either the \code{height} parameter or \code{hash} parameter, but
not both.
}
\examples{
# Look up by height:
# In the following example, block 912345 is looked up by its height.
# Note that block 912345 does not have any non-coinbase transactions.
# (See the next example for a block with extra transactions):
\dontrun{
getBlock(912345)
}

# Look up by hash:
# In the following example, block 993056 is looked up by its hash. Note that
# block 993056 has 3 non-coinbase transactions:
\dontrun{
 getBlock(hash="510ee3c4e14330a7b96e883c323a60ebd1b5556ac1262d0bc03c24a3b785516f")
}
}
\references{
\url{https://getmonero.org/knowledge-base/developer-guides/daemon-rpc#getblock}
}
\author{
Garrett See
}
