% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monero.R
\name{getBlockCount}
\alias{getBlockCount}
\title{getBlockCount}
\usage{
getBlockCount(ip = getOption("monerod.ip", "127.0.0.1"),
  port = getOption("monerod.port", 18081))
}
\arguments{
\item{ip}{daemon ip address}

\item{port}{daemon port}
}
\value{
integer number of blocks in longest chain seen by the node.

numeric
}
\description{
Look up how many blocks are in the longest chain known to the node.
}
\examples{
\dontrun{
getBlockCount()
}
}
\references{
\url{https://getmonero.org/knowledge-base/developer-guides/daemon-rpc#getblockcount}
}
\author{
Garrett See
}
