% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monero.R
\name{getBlockHeaderByHeight}
\alias{getBlockHeaderByHeight}
\title{getBlockHeaderByHeight}
\usage{
getBlockHeaderByHeight(height, ip = getOption("monerod.ip", "127.0.0.1"),
  port = getOption("monerod.port", 18081))
}
\arguments{
\item{height}{unsigned int; The block's height.}

\item{ip}{daemon ip address}

\item{port}{daemon port}
}
\value{
block_header - A structure containing block header information.
  See \code{getLastBlockHeader}.
}
\description{
Similar to \code{getBlockHeaderByHash}, this method includes a
block's height as an input parameter to retrieve basic information about
the block.
}
\examples{
\dontrun{
getBlockHeaderByHeight(912345)
}
}
\references{
\url{https://getmonero.org/knowledge-base/developer-guides/daemon-rpc#getblockheaderbyheight}
}
\author{
Garrett See
}
