% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monero.R
\name{getConnections}
\alias{getConnections}
\title{getConnections}
\usage{
getConnections(ip = getOption("monerod.ip", "127.0.0.1"),
  port = getOption("monerod.port", 18081))
}
\arguments{
\item{ip}{daemon ip address}

\item{port}{daemon port}
}
\value{
connections - List of all connections and their info
\itemize{
 \item avg_download unsigned int; Average bytes of data downloaded by node.
 \item avg_upload unsigned int; Average bytes of data uploaded by node.
 \item current_download unsigned int; Current bytes downloaded by node.
 \item current_upload unsigned int; Current bytes uploaded by node.
 \item incoming boolean; Is the node getting information from your node?
 \item ip string; The node's IP address.
 \item live_time unsigned int
 \item local_ip boolean
 \item localhost boolean
 \item peer_id string; The node's ID on the network.
 \item port stringl The port that the node is using to connect to the network.
 \item recv_count unsigned int
 \item recv_idle_time unsigned int
 \item send_count unsigned int
 \item send_idle_time unsigned int
 \item state string
}
}
\description{
Retrieve information about incoming and outgoing connections to your node.
}
\references{
\url{https://getmonero.org/knowledge-base/developer-guides/daemon-rpc#getconnections}
}
\author{
Garrett See
}
