% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monero.R
\name{getHeight.monerod}
\alias{getHeight.monerod}
\alias{getHeight}
\title{getHeight}
\usage{
getHeight.monerod(ip = getOption("monerod.ip", "127.0.0.1"),
  port = getOption("monerod.port", 18081))

getHeight(ip = getOption("monerod.ip", "127.0.0.1"),
  port = getOption("monerod.port", 18081))
}
\arguments{
\item{ip}{daemon ip address}

\item{port}{daemon port}
}
\value{
height unsigned int; Current length of th elongest chain known to
the daemon.
}
\description{
Get the node's current height
}
\examples{
\dontrun{
getHeight()
}
}
\references{
\url{https://getmonero.org/knowledge-base/developer-guides/daemon-rpc#getheight}
}
\author{
Garrett See
}
