% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monero.R
\name{getInfo}
\alias{getInfo}
\title{getInfo}
\usage{
getInfo(ip = getOption("monerod.ip", "127.0.0.1"),
  port = getOption("monerod.port", 18081))
}
\arguments{
\item{ip}{daemon ip address}

\item{port}{daemon port}
}
\value{
\itemize{
 \item alt_blocks_count - unsigned int; Number of alternative blocks to main chain.
 \item difficulty - unsigned int; Network difficulty (analogous to the strength of the network)
 \item grey_peerlist_size - unsigned int; Grey Peerlist Size
 \item height - unsigned int; Current length of longest chain known to daemon.
 \item incoming_connections_count - unsigned int; Number of peers connected to and pulling from your node.
 \item outgoing_connections_count - unsigned int; Number of peers that you are connected to and getting information from.
 \item status - string; General RPC error code. "OK" means everything looks good.
 \item target - unsigned int; Current target for next proof of work.
 \item target_height - unsigned int; The height of the next block in the chain.
 \item testnet - boolean; States if the node is on the testnet (true) or mainnet (false).
 \item top_block_hash - string; Hash of the highest block in the chain.
 \item tx_count - unsigned int; Total number of non-coinbase transaction in the chain.
 \item tx_pool_siz - unsigned int; Number of transactions that have been broadcast but not included in a block.
 \item white_peerlist_size - unsigned int; White Peerlist Size
}
}
\description{
Retrieve general information about the state of your node and the network.
}
\references{
\url{https://getmonero.org/knowledge-base/developer-guides/daemon-rpc#getinfo}
}
\author{
Garrett See
}
