% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monero.R
\name{getLastBlockHeader}
\alias{getLastBlockHeader}
\title{getLastBlockHeader}
\usage{
getLastBlockHeader(ip = getOption("monerod.ip", "127.0.0.1"),
  port = getOption("monerod.port", "18081"))
}
\arguments{
\item{ip}{daemon ip address}

\item{port}{daemon port}
}
\description{
get block header of the most recent block
}
\examples{
\dontrun{
getLastBlockHeader()
}
}
\references{
\url{https://getmonero.org/knowledge-base/developer-guides/daemon-rpc#getlastblockheader}
}
\author{
Garrett See
}
