% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monero.R
\name{getLastReward}
\alias{getLastReward}
\title{getLastReward}
\usage{
getLastReward(...)
}
\arguments{
\item{...}{arguments to pass to \code{getLastBlockHeader}}
}
\value{
numeric amount of XMR that the reward of the most recent block.
}
\description{
get reward of the most recent block
}
\details{
cals \code{getLastBlockHeader} to get the reward in atomic units, then
divides by 1e12 to return the reward in XMR.
}
\author{
Garrett See
}
