% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monero.R
\name{getNetHash}
\alias{getNetHash}
\title{getNetHash}
\usage{
getNetHash(ip = getOption("monerod.ip", "127.0.0.1"),
  port = getOption("monerod.port", 18081))
}
\arguments{
\item{ip}{daemon ip address}

\item{port}{daemon port}
}
\value{
hashes per second of the network.
}
\description{
get network hash rate estimate by dividing difficulty by 120
}
\examples{
\dontrun{
getNetHash(ip="node.moneroworld.com")
}
}
\author{
Garrett See
}
