% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monero.R
\name{getTransactionPool}
\alias{getTransactionPool}
\title{getTransactionPool}
\usage{
getTransactionPool(ip = getOption("monerod.ip", "127.0.0.1"),
  port = getOption("monerod.port", "18081"))
}
\arguments{
\item{ip}{daemon ip address}

\item{port}{daemon port}
}
\value{
\itemize{
  \item spent_key_images - List of spent output key images:
    \itemize{
      \item id_hash - string; Key image ID hash.
      \item txs_hashes - string list; Key image transaction hashes.
    }
  \item status - string; General RPC error code. "OK" means everything looks good.
  \item transactions - List of transactions in the mempool that have not been included in a block:
    \itemize{
      \item blob_size - unsigned int; The size of the full transaction blob.
      \item fee - unsigned int; The amount of the mining fee included in the transaction, in atomic units.
      \item id_hash - string; The transaction ID hash.
      \item kept_by_block - boolean; We do not accept transactions that timed out before, unless set true.
      \item last_failed_height - unsigned int; If the transaction has previously timed out, this tells at what height that occured.
      \item last_failed_id_hash - string; Like the previous, this tells the previous transaction ID hash.
      \item max_used_block_height - unsigned int; Tells the height of the most recent block with an output used in this transaction.
      \item max_used_block_hash - string; Tells the hash of the most recent block with an output used in this transaction.
      \item receive_time - unsigned int; The Unix time that the transaction was first seen on the network by the node.
      \item tx_json - json string; JSON structure of all information in the transaction:
        \itemize{
          \item version - Transaction version
          \item unlock_time - If not 0, this tells when a transaction output is spendable.
          \item vin - List of inputs into transaction:
            \itemize{
              \item key - The public key of the previous output spent in this transaction.
              \item amount - The amount of the input, in atomic units.
              \item key_offsets - A list of integer offets to the input.
              \item k_image - The key image for the given input
            }
          \item vout - List of outputs from transaction:
          \itemize{
             \item amount - Amount of transaction output, in atomic units.
             \item target - Output destination information:
             \item key - The stealth public key of the receiver. Whoever owns the private key associated with this key controls this transaction output.
         }
         \item extra - Usually called the "transaction ID" but can be used to include any random 32 bytes.
         \item signatures - List of ignatures used in ring signature to hide the true origin of the transaction.
       }
     }
 }
}
\description{
getTransactionPool
}
\author{
Garrett See
}
