% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monero.R
\name{hardForkInfo}
\alias{hardForkInfo}
\title{hardForkInfo}
\usage{
hardForkInfo(ip = getOption("monerod.ip", "127.0.0.1"),
  port = getOption("monerod.port", 18081))
}
\arguments{
\item{ip}{daemon ip address}

\item{port}{daemon port}
}
\description{
hard fork info
}
\details{
Look up information regarding hard fork voting and readiness.
\itemize{
 \item earliest_height - unsigned int; Block height at which hard fork would be enabled if voted in.
 \item enabled - boolean; Tells if hard fork is enforced.
 \item state - unsigned int; Current hard fork state: 0 (There is likely a hard fork), 1 (An update is needed to fork properly), or 2 (Everything looks good).
 \item status - string; General RPC error code. "OK" means everything looks good.
 \item threshold - unsigned int; Minimum percent of votes to trigger hard fork. Default is 80.
 \item version - unsigned int; The major block version for the fork.
 \item votes - unsigned int; Number of votes towards hard fork.
 \item voting - unsigned int; Hard fork voting status.
 \item window - unsigned int; Number of blocks over which current votes are cast. Default is 10080 blocks.
}
}
\examples{
\dontrun{
hardForkInfo()
}
}
\references{
\url{https://getmonero.org/knowledge-base/developer-guides/daemon-rpc#hardforkinfo}
}
\author{
Garrett See
}
