% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moneRo-package.R
\docType{package}
\name{moneRo}
\alias{moneRo}
\alias{moneRo-package}
\alias{moneRo-package}
\title{moneRo package: \R functions to access the monero RPC APIs}
\description{
moneRo package: \R functions to access the monero RPC APIs
}
\references{
\url{https://getmonero.org}
}
\author{
Garrett See \email{gsee@dvchain.co}
}
\keyword{IO}
\keyword{package}
