% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monero.R
\name{onGetBlockHash}
\alias{onGetBlockHash}
\title{onGetBlockHash}
\usage{
onGetBlockHash(height, ip = getOption("monerod.ip", "127.0.0.1"),
  port = getOption("monerod.port", 18081))
}
\arguments{
\item{height}{integer}

\item{ip}{daemon ip address}

\item{port}{daemon port}
}
\value{
character block hash
}
\description{
Look up a block's hash by its height.
}
\examples{
\dontrun{
onGetBlockHash(height=912345)
}
}
\references{
\url{https://getmonero.org/knowledge-base/developer-guides/daemon-rpc#ongetblockhash}
}
\author{
Garrett See
}
